requirements.txt
python-dotenv==0.20.0
bottle==0.12.21
redis-om==0.0.27
nanoid==2.0.0